split notionMock and its internal helpers
