+++ __tests__/data/notion.spec.ts
@@ -1,6 +1,6 @@
 import { beforeAll, beforeEach, describe, expect, it, jest } from '@jest/globals'
 import Notion from '../../src/data/notion'
-import { NotionMock } from '../support/notionHelper'
+import { NotionMock } from '../support/notionMock'
 
 let notionMock: NotionMock
