+++ __tests__/support/notionMock.ts
@@ -1,70 +1,8 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 import { jest } from '@jest/globals'
-import WeekProperties from '../../src/types/weekProperties'
 import { Client, isFullPage } from '@notionhq/client'
-import { GetPageParameters, GetPageResponse, RichTextItemResponse } from '@notionhq/client/build/src/api-endpoints'
-
-const nCheckbox = (checked: boolean) => ({ checkbox: checked })
-const nDate = (start: string) => ({ date: { start } })
-const nNumber = (number: number) => ({ number })
-const nRichText = (text: string): {
-  type: 'rich_text';
-  rich_text: Array<RichTextItemResponse>;
-  id: string;
-} => ({
-  type: 'rich_text',
-  rich_text: [richTextItem(text)],
-  id: 'some-id',
-})
-const nTitle = (title: string): {
-  type: 'title';
-  title: Array<RichTextItemResponse>;
-  id: string;
-} => ({ type: 'title', title: [richTextItem(title)], id: 'some-id' })
-const richTextItem = (text: string): RichTextItemResponse => ({
-  type: 'text',
-  text: {
-    content: text,
-    link: null,
-  },
-  annotations: {
-    bold: false,
-    italic: false,
-    strikethrough: false,
-    underline: false,
-    code: false,
-    color: 'default',
-  },
-  plain_text: text,
-  href: null,
-})
-
-const nUrl = (url: string) => ({ url })
-
-type QueryBody = {
-  database_id: string
-  page_size?: number
-  filter?: {
-    property: string
-    date?: {
-      equals?: string
-      on_or_after?: string
-    }
-    sorts?: {
-      property: string
-      direction: string
-    }[]
-  }
-}
-
-type WeekResponse = {
-  id: string
-  properties: WeekProperties
-}
-
-type WithAuth<P> = P & {
-  auth?: string;
-};
+import { GetPageParameters, GetPageResponse } from '@notionhq/client/build/src/api-endpoints'
+import { QueryBody, WeekResponse, WithAuth, nCheckbox, nDate, nNumber, nRichText, nTitle, nUrl } from './notionHelpers'
 
 export class NotionMock {
   query: jest.MockedFunction<typeof Client.prototype.databases.query> | undefined
@@ -180,5 +118,4 @@ export class NotionMock {
   })
 
   static mockWeeks = (weeks: WeekResponse[] ) => ({ results: weeks })
-
 }
